+++ src/Sql/Sql/ChangeLog.md
@@ -18,10 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added a new cmdlet to refresh external governance status
+    - 'Invoke-AzSqlServerExternalGovernanceStatusRefresh'
 
 ## Version 4.5.0
-* Added a new cmdlet to refresh external governance status
-    - 'Refresh-AzSqlServerExternalGovernanceStatus'
 
 ## Version 4.4.0
 * Fixed identity assignment in `Set-AzSqlDatabase` cmdlet